# docs/generated/ranking_openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Ranking Service",
    "version" : "0.0.1"
  },
...
  "components" : {
    "schemas" : {
      "ErrorResponse" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          }
        }
      },
      "LinkLolAccountRequest" : {
        "type" : "object",
        "properties" : {
          "puuid" : {
            "type" : "string"
          },
          "gameName" : {
            "type" : "string",
            "pattern" : "\\S"
          },
          "tagLine" : {
            "type" : "string",
            "pattern" : "\\S"
          },
          "region" : {
            "$ref" : "#/components/schemas/Region"
          }
        },
        "required" : [ "gameName", "tagLine", "region" ]
      },
...
      "PlayerDTO" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "discordUsername" : {
            "type" : "string"
          },
          "riotAccounts" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/RiotAccountDTO"
            }
          },
          "soloqMMR" : {
            "type" : "integer",
            "format" : "int32"
          },
          "flexMMR" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
      "Region" : {
        "type" : "string",
        "enum" : [ "BR", "EUNE", "EUW", "LAN", "LAS", "NA", "OCE", "RU", "TR", "ME", "JP", "KR", "SEA", "TW", "VN", "PBE" ]
      },
      "RiotAccountDTO" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "puuid" : {
            "type" : "string"
          },
          "gameName" : {
            "type" : "string"
          },
          "tagLine" : {
            "type" : "string"
          },
          "region" : {
            "$ref" : "#/components/schemas/Region"
          },
          "isPrimary" : {
            "type" : "boolean"
          }
        }
      },
      "SetPrimaryAccountRequest" : {
        "type" : "object",
        "required" : [ "puuid" ],
        "properties" : {
          "puuid" : {
            "type" : "string",
            "pattern" : "\\S"
          }
        }
      },
      "UUID" : {
        "type" : "string",
        "format" : "uuid",
        "pattern" : "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
      },
...
    }
  },
  "paths" : {
...
    "/api/v1/ranking/players/links/{gameName}/{tagLine}" : {
      "delete" : {
        "summary" : "Unlink a League of Legends account",
        "operationId" : "unlinkLolAccount",
        "tags" : [ "Player Management" ],
        "parameters" : [ {
          "name" : "gameName",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "pattern" : "\\S"
          }
        }, {
          "name" : "tagLine",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "pattern" : "\\S"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Account successfully unlinked or not found"
          }
        }
      }
    },
    "/api/v1/ranking/players/{playerId}" : {
      "get" : {
        "summary" : "Get player details",
        "operationId" : "getPlayer",
        "tags" : [ "Player Management" ],
        "parameters" : [ {
          "description" : "Player UUID",
          "example" : "550e8400-e29b-41d4-a716-446655440002",
          "name" : "playerId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Player found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PlayerDTO"
                }
              }
            }
          },
          "404" : {
            "description" : "Player not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ranking/players/{playerId}/link-lol-account" : {
      "put" : {
        "summary" : "Link a League of Legends account to a player",
        "description" : "For gameName/tagLine discovery, you can use https://www.leagueofgraphs.com/rankings/summoners . Region should be the Riot routing region enum (AMERICAS, EUROPE, ASIA, SEA).",
        "operationId" : "linkLolAccount",
        "tags" : [ "Player Management" ],
        "parameters" : [ {
          "name" : "playerId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/LinkLolAccountRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Account successfully linked",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PlayerDTO"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid account info or already linked",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "Player not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ranking/players/{playerId}/primary-account" : {
      "get" : {
        "summary" : "Get the primary linked Riot account for a player",
        "operationId" : "getPrimaryAccount",
        "tags" : [ "Player Management" ],
        "parameters" : [ {
          "description" : "Player UUID",
          "example" : "550e8400-e29b-41d4-a716-446655440002",
          "name" : "playerId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Primary account found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RiotAccountDTO"
                }
              }
            }
          },
          "400" : {
            "description" : "No League of Legends account linked",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "Player not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/ranking/players/{playerId}/set-primary-account" : {
      "put" : {
        "summary" : "Set one of the linked Riot accounts as the primary account",
        "operationId" : "setPrimaryAccount",
        "tags" : [ "Player Management" ],
        "parameters" : [ {
          "name" : "playerId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SetPrimaryAccountRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Primary account successfully updated",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PlayerDTO"
                }
              }
            }
          },
          "404" : {
            "description" : "Player or Riot account not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          }
        }
      }
    },
    "/api/v1/ranking/players/{playerId}/sync-mmr" : {
      "post" : {
        "summary" : "Manually fetch and sync the player's MMR from the external LoL service",
        "operationId" : "syncExternalMMR",
        "tags" : [ "Player Management" ],
        "parameters" : [ {
          "name" : "playerId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "MMR successfully synced",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PlayerDTO"
                }
              }
            }
          },
          "400" : {
            "description" : "League of Legends account not linked",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "Player not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Auto generated value"
  }, {
    "url" : "http://0.0.0.0:8080",
    "description" : "Auto generated value"
  } ]
}